refactor: 가격 표시 범위를 Long -> BigDecimal 로 수정

+++ core/designsystem/src/main/java/com/susu/core/designsystem/component/textfield/PriceVisualTransformation.kt
@@ -9,7 +9,7 @@ import java.text.DecimalFormat
 class PriceVisualTransformation : VisualTransformation {
     override fun filter(text: AnnotatedString): TransformedText {
         val amount = text.text
-        val formatAmount = if (amount.isEmpty()) "" else DecimalFormat("#,###").format(amount.toLong())
+        val formatAmount = if (amount.isEmpty()) "" else DecimalFormat("#,###").format(amount.toBigDecimal())
 
         return TransformedText(
             text = AnnotatedString("${formatAmount}원"),
